refactor(feg): Segregation of Accounting servicer as external (#10998)
